don't use new auth in production
